Move the request handler to SchemaHandler
